time_manager/backend/requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiogram-calendar==0.5.0
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.28.0
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.0.0.post2
fastapi==0.109.2
fastapi-cache2==0.2.1
frozenlist==1.4.0
fuzzywuzzy==0.18.0
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Levenshtein==0.23.0
magic-filter==1.0.12
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
orjson==3.9.9
pendulum==3.0.0
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.23.0
python-multipart==0.0.6
python-telegram-bot-calendar==1.0.5
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.4.0
redis==4.6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.36.3
time-machine==2.13.0
typing_extensions==4.8.0
tzdata==2024.1
ujson==5.8.0
urllib3==2.1.0
uvicorn==0.27.1
watchfiles==0.21.0
websockets==11.0.3
yarl==1.9.3
